docs: extract react-live-codeblock module
